# oapispec/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the NBRGLM Nexeres API Spec documentation.",
        "title": "NBRGLM Nexeres API Spec",
        "termsOfService": "https://nbrglm.com/nexeres/terms",
        "contact": {
            "name": "NBRGLM Support",
            "url": "https://nbrglm.com/support",
            "email": "contact@nbrglm.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.0.1"
    },
    "host": "localhost:3360",
    "basePath": "/",
    "paths": {
...
        "/api/auth/login": {
            "post": {
                "description": "Handles user login requests.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "User Login",
                "parameters": [
                    {
                        "description": "User Login Data",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UserLoginData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Login Result",
                        "schema": {
                            "$ref": "#/definitions/handlers.UserLoginResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid Credentials or User does not belong to any organization",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "description": "Logs out the user by revoking their session using session token or refresh token. Requires atleast one of the tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Logout user",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session token",
                        "name": "X-NEXERES-Session-Token",
                        "in": "header"
                    },
                    {
                        "type": "string",
                        "description": "Refresh token",
                        "name": "X-NEXERES-Refresh-Token",
                        "in": "header"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Logout result",
                        "schema": {
                            "$ref": "#/definitions/handlers.LogoutResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid or missing tokens",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid or expired tokens",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.LogoutResult": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "handlers.UserLoginData": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "flowReturnTo": {
                    "description": "Optional field to store in the flow data which can be fetched by the client after login\nThis can be used to redirect the user to a specific page after login\nor to maintain the state of the application.\nIt is recommended to validate this field on the client side to prevent open redirect vulnerabilities.",
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "handlers.UserLoginResult": {
            "type": "object",
            "properties": {
                "flowId": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "requireEmailVerification": {
                    "type": "boolean"
                },
                "tokens": {
                    "$ref": "#/definitions/tokens.Tokens"
                }
            }
        },
...
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "debug": {
                    "description": "DebugMessage is a technical message that can be used for debugging.",
                    "type": "string"
                },
                "message": {
                    "description": "Message is a user-friendly message that can be displayed to the end user.",
                    "type": "string"
                }
            }
        },
...
        "tokens.Tokens": {
            "type": "object",
            "properties": {
                "refreshToken": {
                    "description": "RefreshToken is the generated refresh token.\n\nThis is base64.RawURLEncoding encoded.\nDO NOT DECODE IT WHILE RETRIEVING IT FROM THE COOKIES/CLIENT.",
                    "type": "string"
                },
                "refreshTokenExpiry": {
                    "description": "RefreshTokenExpiry is the expiration time of the refresh token.",
                    "type": "string"
                },
                "sessionId": {
                    "description": "SessionId is the unique identifier for the session.",
                    "type": "string"
                },
                "sessionToken": {
                    "description": "SessionToken is the generated session token.\n\nThis is a jwt which is base64.RawURLEncoding encoded.\nYOU NEED TO DECODE IT WHILE RETRIEVING IT FROM THE COOKIES/CLIENT.\nDO NOT USE IT AS IS. VALIDATION WILL FAIL WITHOUT DECODING.\nONLY WHEN DECODED, YOU SHOULD PASS IT TO THE THINGS THAT REQUIRE THE SESSION TOKEN.",
                    "type": "string"
                },
                "sessionTokenExpiry": {
                    "description": "SessionTokenExpiry is the expiration time of the session token.",
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "APIKeyAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-API-Key",
            "in": "header"
        },
        "AdminHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Admin-Token",
            "in": "header"
        },
        "RefreshHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Refresh-Token",
            "in": "header"
        },
        "SessionHeaderAuth": {
            "type": "apiKey",
            "name": "X-NEXERES-Session-Token",
            "in": "header"
        }
    },
...
}